AI_logic/pyproject.toml
[project]
name = "ai-logic"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<4"
dependencies = [
    "anthropic (==0.40.0)",
    "cohere (==5.5.3)",
    "tiktoken (==0.7.0)",
    "mistralai (==1.9.10)",
    "chromadb (==0.5.0)",
    "numpy (==1.26.4)",
    "sympy (==1.12)",
    "python-dotenv (==1.0.1)",
    "pydantic (==2.10.3)",
    "httpx (==0.28.1)",
    "rich (==13.7.1)",
    "pytest (==8.2.0)",
    "langchain-community (==0.0.38)",
    "langchain-text-splitters (==0.0.1)",
    "pymupdf (==1.24.4)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
